Use proper HTTP status code 204 if no mail server is configured
Before, the status code 404 (not found) has been returned if the user tried to
get the configuration of a mail server that has not been configured. This is a
problem as 404 is also returned if the REST endpoint is not available, which is
also the case if the plugin cannot be loaded.

Instead, now the status code 204 (no content) is used. It is unusual to use it
for a GET request but makes sense to distinguish between 200 (an actual
configuration has been returned) and 204 (the request was successful but there
is no mail configuration yet).