Improve swagger usage, switch to asciidoc
In order to be more consistent, the dependencies in the pom.xml have been
slightly adapted. The generated swagger.yaml file in the project root has been
deleted for now as it should not be used yet for anything else then
documentation generation. The problem is that the order of the paths is randomly
changed from time to time and thus generated useless changes in Git.

Furthermore, the generated file format has been changed from markdown to
asciidoc. Asciidoc is considered more stable, generated the whole documentation
into a single file (well, almost) and allows including further, manually
maintained documentation or (autogenerated) JSON example files. Unfortunately,
this include feature of asciidoc is not supported by GitHub, but this won't be a
problem anymore when we deploy the documentation somewhere else.