Reduce swagger annotations, use 'parent' openapi.yaml instead
The class-wide OpenAPIDefinition annotation does not behave as expected.
Instead of providing a documentation for only this resource, the documentation
is considered as global documentation for the whole plugin API and thus has been
removed in favor of the introduced manually maintained openapi.yaml file.

A lot of descriptions have been removed as they (in most cases) provide a very
small value but interfere with the readability of the code and result into
lengths of the particular code lines that go far beyond the recommended length
of 120 characters.

The same goes for the Parameter annotations. Their descriptions in most cases
provide very low value but result in very high repetetions (thus violate DRY,
e.g. Content(schema = Schema(implementation = UserDirectoryBean.class)))
UserDirectoryBean dir). Required and not required parameters should be annotated
with NotNull (?) and Nullable annotations instead, the names should document
itself.

Regarding the last point:

The 'test' query parameter also violated the principle of least astonishment
(PolA) as a null (not given) value was considered as true
(Boolean.parseValue(null) e.g. evaluates to false). In order to get rid of the
additional documentation, the parameter was renamed to testConnection and per
default, the connection is not tested anymore.