Update Swagger deps, use Maven wrapper, always update deps
The used Swagger dependency contains classes that caused problems in the Crowd
ConfAPI plugin when trying to compile it into the plugin as Swagger is not a
proviced dependency in Crowd. This is why the dependency has been changed here
too and is now also compiled into the plugin (it only contains annotations).

In order to fix the IT soon, the Maven wrapper has been integrated with a fixed
Maven version of 3.5.4 as this is also the Maven version used by Atlassian
currently.

The Circle CI pipeline has been updated to use that wrapper. Maven dependency
metadata will now always be updated to notice changes in cached SNAPSHOT
versions.