Adjust initial user resource implementation to latest standards
Generic files like the user bean, the user service implementation and the user
resource implementation have been moved to the ConfAPI commons library while the
actual service implementation is kept at the plugin itself.

In order to have a 'centralized' translation of user and user beans, a new user
bean util has been introduced. The service and the resource as possible
alternatives to this util turned out to be hard to handle.