Use swagger annotated resource interface (PoC)
It is possible to move the swagger annotation to resource interfaces. This has
the benefit that all features (annotations) of swagger (that make the code very
hard to read) can be used apart from the actual implementation but still in the
code. The other large benefit is that the same documentation can be shared
amongst all plugins (if needed).