Abstract integration tests implementation
This commit introduces implementation for abstract
integration testing based on confapi-commons.
Examples are given for 'ping', 'settings' and
'licenses' resources.

The intention is to define all integration tests in the
commons library and only inherit the 'Abstract*ResourceFuncTest'
classes in the actual confapi plugin.

Throughout the execution of the licenses integration test a bug was
identified and fixed which arised when setting multiple licenses at
once and thus, removing existing licenses first. For Confluence one
has to check if multi licenses are supported on the platform at all.