Space permissions integration tests implementation
This commit implements space permissions integration tests. Note that these
tests have no abstract counterpart in the commons-lib because other Atlassian
products do not share the ability to manage spaces.

In accordance with the current service interface achitecture implemented in the
commons lib the permissions resource was refactored to represent permission
interface and resource consumer as well. Thus, this implementation is future proof
when further abstraction might be necessary.

Finally, a bean example wor permission access was added for generic test use.